Both sync.Map and cowmap use atomic.Value atomic operations to access the map during data reading, resulting in similar read performance. However, sync.Map has better write performance. Therefore, cowmap directly utilizes sync.Map as its internal structure.

+++ eventbus_test.go
@@ -81,7 +81,7 @@ func Test_channelClose(t *testing.T) {
 	err := ch.subscribe(busHandlerOne)
 	assert.Nil(t, err)
 	ch.close()
-	assert.Equal(t, 0, ch.handlers.Len())
+	assert.Equal(t, uint32(0), ch.handlers.Len())
 	ch.close()
 }
